// pro/src/apiClient/v1/models/ThingStockCreateBodyModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ThingStockCreateBodyModel = {
  activationCodes?: Array<string> | null;
  activationCodesExpirationDatetime?: string | null;
  bookingLimitDatetime?: string | null;
  offerId: number;
  price: number;
  quantity?: number | null;
};